ex: 6.2

+++ shaders.cpp
@@ -78,5 +78,5 @@ void Shader::setInt(const std::string& name, int value) const {
 }
 
 void Shader::setFloat(const std::string& name, float value) const {
-  glUniform1f(glGetUniformLocation(ID, name.c_str()), (int)value);
+  glUniform1f(glGetUniformLocation(ID, name.c_str()), value);
 }
